+++ src/main/java/org/opentosca/container/client/model/Interface.java
@@ -4,20 +4,18 @@ import java.util.List;
 import java.util.stream.Collectors;
 
 import io.swagger.client.model.InterfaceDTO;
-import io.swagger.client.model.PlanDTO;
 import lombok.RequiredArgsConstructor;
 
 @RequiredArgsConstructor
 public class Interface {
 
-	private final InterfaceDTO iface;
-	
-	public String getName() {
-		return this.iface.getName();
-	}
-	
-	public List<Operation> getOperations() {
-		return this.iface.getOperations().values().stream().map(x -> new Operation(x)).collect(Collectors.toList());
-	}
-	
+    private final InterfaceDTO iface;
+
+    public String getName() {
+        return this.iface.getName();
+    }
+
+    public List<Operation> getOperations() {
+        return this.iface.getOperations().values().stream().map(x -> new Operation(x)).collect(Collectors.toList());
+    }
 }
