+++ src/test/java/org/opentosca/container/client/SmartServiceContainerClientTests.java
@@ -7,38 +7,31 @@ import java.util.List;
 import java.util.Map;
 
 import io.swagger.client.model.ServiceTemplateInstanceDTO;
-import lombok.Getter;
-import lombok.Setter;
-
 import org.json.JSONObject;
 import org.junit.Assert;
 import org.junit.Assume;
 import org.junit.Before;
 import org.junit.FixMethodOrder;
-import org.junit.Ignore;
 import org.junit.Test;
 import org.junit.runner.RunWith;
 import org.junit.runners.MethodSorters;
 import org.opentosca.container.client.impl.SmartServiceSwaggerContainerClient.NoSmartServiceException;
-import org.opentosca.container.client.model.*;
+import org.opentosca.container.client.model.Application;
+import org.opentosca.container.client.model.ApplicationInstance;
+import org.opentosca.container.client.model.Plan;
+import org.opentosca.container.client.model.PlanInstance;
+import org.opentosca.container.client.model.PlanType;
 import org.springframework.beans.factory.annotation.Autowired;
-import org.springframework.boot.SpringApplication;
-import org.springframework.boot.autoconfigure.SpringBootApplication;
-import org.springframework.boot.context.properties.ConfigurationProperties;
-import org.springframework.boot.context.properties.EnableConfigurationProperties;
 import org.springframework.boot.test.context.SpringBootTest;
 import org.springframework.test.context.junit4.SpringRunner;
 
-@Ignore
 @RunWith(SpringRunner.class)
 @SpringBootTest
-//@SpringBootApplication
-@EnableConfigurationProperties(SmartServiceContainerClientTests.Config.class)
 @FixMethodOrder(MethodSorters.NAME_ASCENDING)
 public class SmartServiceContainerClientTests {
 
     @Autowired
-    private SmartServiceContainerClientTests.Config config;
+    private ClientTests.Config config;
 
     private SmartServiceContainerClient client;
 
@@ -66,17 +59,17 @@ public class SmartServiceContainerClientTests {
 
     @Test
     public void test_20_upload() {
-        for (Config.Test test : config.getTests()) {
+        for (ClientTests.Config.Test test : config.getTests()) {
             Assert.assertFalse(client.getApplication(test.getName()).isPresent());
             Path path = Paths.get(config.getPath(), test.getName());
-            Application application = client.uploadApplication(path);            
+            Application application = client.uploadApplication(path);
             Assert.assertEquals(test.getName(), application.getId());
             try {
-				JSONObject smartServiceDescription = client.getSmartServiceDescription(application);
-				Assert.assertNotNull(smartServiceDescription);
-			} catch (NoSmartServiceException e) {
-				Assert.fail("Application under test is not a smart service");
-			}
+                JSONObject smartServiceDescription = client.getSmartServiceDescription(application);
+                Assert.assertNotNull(smartServiceDescription);
+            } catch (NoSmartServiceException e) {
+                Assert.fail("Application under test is not a smart service");
+            }
         }
         List<Application> applications = client.getApplications();
         Assert.assertEquals(config.getTests().size(), applications.size());
@@ -84,7 +77,7 @@ public class SmartServiceContainerClientTests {
 
     @Test
     public void test_30_provision_application() {
-        for (Config.Test test : config.getTests()) {
+        for (ClientTests.Config.Test test : config.getTests()) {
             Application application = client.getApplication(test.getName()).orElseThrow(IllegalStateException::new);
             Assert.assertEquals(0, client.getApplicationInstances(application, ServiceTemplateInstanceDTO.StateEnum.CREATED).size());
             int startSize = client.getApplicationInstances(application).size();
@@ -98,7 +91,7 @@ public class SmartServiceContainerClientTests {
 
     @Test
     public void test_40_get_application_instances() {
-        for (Config.Test test : config.getTests()) {
+        for (ClientTests.Config.Test test : config.getTests()) {
             Application application = client.getApplication(test.getName()).orElseThrow(IllegalStateException::new);
             List<ApplicationInstance> applicationInstances = client.getApplicationInstances(application, ServiceTemplateInstanceDTO.StateEnum.CREATED);
             Assert.assertEquals(1, applicationInstances.size());
@@ -122,7 +115,7 @@ public class SmartServiceContainerClientTests {
 
     @Test
     public void test_45_execute_node_operation() {
-        for (Config.Test test : config.getTests()) {
+        for (ClientTests.Config.Test test : config.getTests()) {
             Application application = client.getApplication(test.getName()).orElseThrow(IllegalStateException::new);
             List<ApplicationInstance> applicationInstances = client.getApplicationInstances(application, ServiceTemplateInstanceDTO.StateEnum.CREATED);
             Assert.assertEquals(1, applicationInstances.size());
@@ -144,7 +137,7 @@ public class SmartServiceContainerClientTests {
 
     @Test
     public void test_50_get_buildplan_instances() {
-        for (Config.Test test : config.getTests()) {
+        for (ClientTests.Config.Test test : config.getTests()) {
             Application application = client.getApplication(test.getName()).orElseThrow(IllegalStateException::new);
             List<PlanInstance> buildPlanInstances = application.getBuildPlanInstances();
             Assert.assertNotNull(buildPlanInstances);
@@ -154,7 +147,7 @@ public class SmartServiceContainerClientTests {
 
     @Test
     public void test_60_terminate_instance() {
-        for (Config.Test test : config.getTests()) {
+        for (ClientTests.Config.Test test : config.getTests()) {
             Application application = client.getApplication(test.getName()).orElseThrow(IllegalStateException::new);
             List<ApplicationInstance> applicationInstances = client.getApplicationInstances(application, ServiceTemplateInstanceDTO.StateEnum.CREATED);
             Assert.assertEquals(1, applicationInstances.size());
@@ -170,27 +163,5 @@ public class SmartServiceContainerClientTests {
             Assert.assertTrue(client.removeApplication(application));
         }
         Assert.assertEquals(0, client.getApplications().size());
-    }
-
-    public static void main(String[] args) {
-        SpringApplication.run(SmartServiceContainerClientTests.class, args);
-    }
-
-    @Setter
-    @Getter
-    @ConfigurationProperties(prefix = "csar")
-    static class Config {
-
-        private String hostname;
-        private String path;
-        private List<Test> tests;
-
-        @Setter
-        @Getter
-        static class Test {
-
-            private String name;
-            private Map<String, String> input;
-        }
     }
 }
