+++ src/test/resources/application.yml
@@ -2,7 +2,7 @@ csar:
   hostname: localhost
   path: c:/temp
   tests:
-  - name: MyTinyToDo_Bare_Docker.csar
-    input:
-      DockerEngineURL: tcp://dind:2375
-      ApplicationPort: 9990
+    - name: MyTinyToDo_Bare_Docker.csar
+      input:
+        DockerEngineURL: tcp://dind:2375
+        ApplicationPort: 9990
