+++ src/test/resources/application2.yml
@@ -2,9 +2,9 @@ csar:
   hostname: localhost
   path: c:/temp
   tests:
-  - name: UpdateablePythonRuleService_ServiceClass2_SmartOrchestra_WithoutCloud_w1-wip1.csar
-    input:
-      VMUserName: ubuntu
-      VMPrivateKey: --BEGINRSA--asdqwe
-      VMIP: 192.168.2.1
-      RuleSource: import time; \n\ndef main():\n\twhile True:\n\t\tprint ("HelloWorld")\nif __name__ == "__main__":\n\tmain()
+    - name: UpdateablePythonRuleService_ServiceClass2_SmartOrchestra_WithoutCloud_w1-wip1.csar
+      input:
+        VMUserName: ubuntu
+        VMPrivateKey: --BEGINRSA--asdqwe
+        VMIP: 192.168.2.1
+        RuleSource: import time; \n\ndef main():\n\twhile True:\n\t\tprint ("HelloWorld")\nif __name__ == "__main__":\n\tmain()
